Fix signedness warning in glib/test/tree.c
glib/tests/tree.c: In function ‘test_tree_traverse’:
glib/tests/tree.c:394:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’
  394 |   for (i = 0; i < G_N_ELEMENTS (orders); i++)
      |                 ^

+++ glib/tests/scannerapi.c
@@ -111,8 +111,8 @@ test_scanner_tokens (ScannerFixture *fix,
   gchar buf[] = "(\t\n\r\\){}";
   const gint buflen = strlen (buf);
   gchar tokbuf[] = "(\\){}";
-  const gsize tokbuflen = strlen (tokbuf);
-  gsize i;
+  const gint tokbuflen = strlen (tokbuf);
+  guint i;
 
   g_scanner_input_text (fix->scanner, buf, buflen);
